PARALLEL SEARCH/backend/requirements.txt
Flask
flask-cors
pdfplumber
pdfminer.six
fpdf
python-dotenv
pymupdf
python-docx 
PyPDF2
